01_spark_blockchain_pipeline/requirements.txt
pyspark
python-dotenv
pandas
numpy
psycopg2-binary
requests
websocket-client
websockets